change how L3Outs are created
On external network creation, always create a L3Out on the right
VRF (may or may not be a NAT vrf).

When a router port is attached as GW, create a shadow L3Out in the
proper VRF when NAT is involved, connect the contracts following
the old workflow otherwise.